Musical Dice Game
This is a project i created during a college course and it will compose a two-part waltz by pasting together 32 of 272 per-composed musical elements at random. requires simpleaudio python package to run. The waltz consists of two parts - the minuet and the trio. Each is comprised of 16 measures, which are generated at random. The minuet consists of 16 measures. There are 176 possible Minuet measures, named M1.wav through M176.wav. The trio consists of 16 measures. There are 96 possible Trio measures named T1.wav through T96.wav. To determine which measure to play 2 dice are rolled. There are tables that determine which measure is played but they are not represented here.